refactors bindings
