+++ ConnectorAutocadCivil/ConnectorAutocadCivil/UI/ConnectorBindingsAutocadCivil.Send.cs
@@ -1,0 +1,306 @@
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Autodesk.AutoCAD.ApplicationServices;
+using Autodesk.AutoCAD.DatabaseServices;
+using DesktopUI2;
+using DesktopUI2.Models;
+using DesktopUI2.ViewModels;
+using Speckle.Core.Api;
+using Speckle.Core.Kits;
+using Speckle.Core.Logging;
+using Speckle.Core.Models;
+using Speckle.Core.Transports;
+using static Speckle.ConnectorAutocadCivil.Utils;
+
+#if ADVANCESTEEL2023
+using ASFilerObject = Autodesk.AdvanceSteel.CADAccess.FilerObject;
+#endif
+
+namespace Speckle.ConnectorAutocadCivil.UI
+{
+  public partial class ConnectorBindingsAutocad : ConnectorBindings
+  {
+    public override async Task<string> SendStream(StreamState state, ProgressViewModel progress)
+    {
+      var converter = KitManager.GetDefaultKit().LoadConverter(Utils.VersionedAppName);
+
+      var streamId = state.StreamId;
+      var client = state.Client;
+      progress.Report = new ProgressReport();
+
+      if (state.Filter != null)
+        state.SelectedObjectIds = GetObjectsFromFilter(state.Filter, converter);
+
+      // remove deleted object ids
+      var deletedElements = new List<string>();
+      foreach (var selectedId in state.SelectedObjectIds)
+        if (Utils.GetHandle(selectedId, out Handle handle))
+          if (Doc.Database.TryGetObjectId(handle, out ObjectId id))
+            if (id.IsErased || id.IsNull)
+              deletedElements.Add(selectedId);
+
+      state.SelectedObjectIds = state.SelectedObjectIds.Where(o => !deletedElements.Contains(o)).ToList();
+
+      if (state.SelectedObjectIds.Count == 0)
+      {
+        throw new InvalidOperationException(
+          "Zero objects selected; send stopped. Please select some objects, or check that your filter can actually select something."
+        );
+      }
+
+      var commitObject = new Base();
+      commitObject["units"] = Utils.GetUnits(Doc); // TODO: check whether commits base needs units attached
+
+      int convertedCount = 0;
+
+      // invoke conversions on the main thread via control
+      try
+      {
+        if (Control.InvokeRequired)
+          Control.Invoke(
+            new Action(() => ConvertSendCommit(commitObject, converter, state, progress, ref convertedCount)),
+            new object[] { }
+          );
+        else
+          ConvertSendCommit(commitObject, converter, state, progress, ref convertedCount);
+        progress.Report.Merge(converter.Report);
+      }
+      catch (Exception e)
+      {
+        progress.Report.LogOperationError(e);
+      }
+
+      if (convertedCount == 0)
+      {
+        throw new SpeckleException("Zero objects converted successfully. Send stopped.");
+      }
+
+      progress.CancellationToken.ThrowIfCancellationRequested();
+
+      var transports = new List<ITransport>() { new ServerTransport(client.Account, streamId) };
+
+      var commitObjId = await Operations.Send(
+        commitObject,
+        progress.CancellationToken,
+        transports,
+        onProgressAction: dict => progress.Update(dict),
+        onErrorAction: ConnectorHelpers.DefaultSendErrorHandler,
+        disposeTransports: true
+      );
+
+      progress.CancellationToken.ThrowIfCancellationRequested();
+
+      var actualCommit = new CommitCreateInput
+      {
+        streamId = streamId,
+        objectId = commitObjId,
+        branchName = state.BranchName,
+        message = state.CommitMessage ?? $"Pushed {convertedCount} elements from {Utils.AppName}.",
+        sourceApplication = Utils.VersionedAppName
+      };
+
+      if (state.PreviousCommitId != null)
+      {
+        actualCommit.parents = new List<string>() { state.PreviousCommitId };
+      }
+
+      var commitId = await ConnectorHelpers.CreateCommit(client, actualCommit, progress.CancellationToken);
+      return commitId;
+    }
+
+    delegate void SendingDelegate(
+      Base commitObject,
+      ISpeckleConverter converter,
+      StreamState state,
+      ProgressViewModel progress,
+      ref int convertedCount
+    );
+
+    private void ConvertSendCommit(
+      Base commitObject,
+      ISpeckleConverter converter,
+      StreamState state,
+      ProgressViewModel progress,
+      ref int convertedCount
+    )
+    {
+      using (DocumentLock acLckDoc = Doc.LockDocument())
+      {
+        using (Transaction tr = Doc.Database.TransactionManager.StartTransaction())
+        {
+          // set the context doc for conversion - this is set inside the transaction loop because the converter retrieves this transaction for all db editing when the context doc is set!
+          converter.SetContextDocument(Doc);
+
+          // set converter settings as tuples (setting slug, setting selection)
+          var settings = new Dictionary<string, string>();
+          CurrentSettings = state.Settings;
+          foreach (var setting in state.Settings)
+            settings.Add(setting.Slug, setting.Selection);
+          converter.SetConverterSettings(settings);
+
+          var conversionProgressDict = new ConcurrentDictionary<string, int>();
+          conversionProgressDict["Conversion"] = 0;
+
+          // add applicationID xdata before send
+          if (!ApplicationIdManager.AddApplicationIdXDataToDoc(Doc, tr))
+          {
+            progress.Report.LogOperationError(new Exception("Could not create document application id reg table"));
+            return;
+          }
+
+          // get the hash of the file name to create a more unique application id
+          var fileNameHash = GetDocumentId();
+
+          string servicedApplication = converter.GetServicedApplications().First();
+
+          foreach (var autocadObjectHandle in state.SelectedObjectIds)
+          {
+            // handle user cancellation
+            if (progress.CancellationToken.IsCancellationRequested)
+            {
+              return;
+            }
+
+            // get the db object from id
+            DBObject obj = null;
+            string layer = null;
+            string applicationId = null;
+            if (Utils.GetHandle(autocadObjectHandle, out Handle hn))
+            {
+              obj = hn.GetObject(tr, out string type, out layer, out applicationId);
+            }
+            else
+            {
+              progress.Report.LogOperationError(new Exception($"Failed to find doc object ${autocadObjectHandle}."));
+              continue;
+            }
+
+            // create applicationobject for reporting
+            Base converted = null;
+            var descriptor = Utils.ObjectDescriptor(obj);
+            ApplicationObject reportObj = new ApplicationObject(autocadObjectHandle, descriptor)
+            {
+              applicationId = autocadObjectHandle
+            };
+
+            if (!converter.CanConvertToSpeckle(obj))
+            {
+#if ADVANCESTEEL2023
+              UpdateASObject(reportObj, obj);
+#endif
+              reportObj.Update(
+                status: ApplicationObject.State.Skipped,
+                logItem: $"Sending this object type is not supported in {Utils.AppName}"
+              );
+              progress.Report.Log(reportObj);
+              continue;
+            }
+
+            try
+            {
+              // convert obj
+              converter.Report.Log(reportObj); // Log object so converter can access
+              converted = converter.ConvertToSpeckle(obj);
+              if (converted == null)
+              {
+                reportObj.Update(status: ApplicationObject.State.Failed, logItem: $"Conversion returned null");
+                progress.Report.Log(reportObj);
+                continue;
+              }
+
+              /* TODO: adding the extension dictionary / xdata per object
+              foreach (var key in obj.ExtensionDictionary)
+                converted[key] = obj.ExtensionDictionary.GetUserString(key);
+              */
+
+#if CIVIL2021 || CIVIL2022 || CIVIL2023 || CIVIL2024
+              // add property sets if this is Civil3D
+              var propertySets = obj.GetPropertySets(tr);
+              if (propertySets.Count > 0)
+                converted["propertySets"] = propertySets;
+#endif
+
+              string containerName = obj is BlockReference ? "Blocks" : Utils.RemoveInvalidDynamicPropChars(layer); // remove invalid chars from layer name
+
+              if (commitObject[$"@{containerName}"] == null)
+                commitObject[$"@{containerName}"] = new List<Base>();
+              ((List<Base>)commitObject[$"@{containerName}"]).Add(converted);
+
+              // set application id
+              #region backwards compatibility
+              // this is just to overwrite old files with objects that have the unappended autocad native application id
+              bool isOldApplicationId(string appId)
+              {
+                if (string.IsNullOrEmpty(appId))
+                  return false;
+                return appId.Length == 5 ? true : false;
+              }
+              if (isOldApplicationId(applicationId))
+              {
+                ApplicationIdManager.SetObjectCustomApplicationId(
+                  obj,
+                  autocadObjectHandle,
+                  out applicationId,
+                  fileNameHash
+                );
+              }
+              #endregion
+
+              if (applicationId == null) // this object didn't have an xdata appId field
+              {
+                if (
+                  !ApplicationIdManager.SetObjectCustomApplicationId(
+                    obj,
+                    autocadObjectHandle,
+                    out applicationId,
+                    fileNameHash
+                  )
+                )
+                {
+                  reportObj.Log.Add("Could not set application id xdata");
+                }
+              }
+              converted.applicationId = applicationId;
+
+              // update progress
+              conversionProgressDict["Conversion"]++;
+              progress.Update(conversionProgressDict);
+
+              // log report object
+              reportObj.Update(status: ApplicationObject.State.Created, logItem: $"Sent as {converted.GetType().Name}");
+              progress.Report.Log(reportObj);
+
+              convertedCount++;
+            }
+            catch (Exception e)
+            {
+              //TODO: Log to serilog failed conversions
+              reportObj.Update(status: ApplicationObject.State.Failed, logItem: $"{e.Message}");
+              progress.Report.Log(reportObj);
+              continue;
+            }
+          }
+
+          tr.Commit();
+        }
+      }
+    }
+
+#if ADVANCESTEEL2023
+    private void UpdateASObject(ApplicationObject applicationObject, DBObject obj)
+    {
+      if (!CheckAdvanceSteelObject(obj))
+        return;
+
+      ASFilerObject filerObject = GetFilerObjectByEntity<ASFilerObject>(obj);
+      if (filerObject != null)
+      {
+        applicationObject.Update(descriptor: filerObject.GetType().Name);
+      }
+    }
+#endif
+  }
+}
